refactor(dispatch): extract DispatchManifest + RiderInfo into manifest.rs
Pure file-move refactor — no behavior change. The manifest is the
"per-tick demand picture" handed to dispatch strategies; it's a
self-contained module with its own data shape and accessor
surface, so giving it a dedicated file shrinks dispatch/mod.rs
(1635 → 1458 LOC) and makes future borrow-vs-clone refactors
easier to scope.

The pub re-exports in dispatch/mod.rs preserve the public API
surface — every existing crate::dispatch::DispatchManifest /
RiderInfo import keeps working unchanged.

First step toward the audit's full split (§20). Future PRs can
extract assignment.rs, group.rs, and strategy.rs as separate
focused refactors.
